Add default enum argument support. Null support for enum in BeanMapper
Split BuiltInArgumentFactory out of Foreman so that BeanMapperFactory can reference it and not offer bean mapping for built-in argument types

Corrected EnumArgument type declaration per comment by stevenschlansker.

Add missing license headers.
