+++ src/main/java/org/skife/jdbi/v2/BeanMapper.java
@@ -118,7 +118,8 @@ public class BeanMapper<T> implements ResultSetMapper<T>
                     value = rs.getString(i);
                 }
                 else if (type.isEnum()) {
-                    value = Enum.valueOf(type, rs.getString(i));
+                    String str = rs.getString(i);
+                    value = str != null ? Enum.valueOf(type, str) : null;
                 }
                 else {
                     value = rs.getObject(i);
