+++ src/main/java/org/skife/jdbi/v2/EnumArgument.java
@@ -1,0 +1,46 @@
+/*
+ * Copyright (C) 2004 - 2014 Brian McCallister
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.skife.jdbi.v2;
+
+import org.skife.jdbi.v2.tweak.Argument;
+
+import java.sql.PreparedStatement;
+import java.sql.SQLException;
+import java.sql.Types;
+
+
+class EnumArgument<T extends Enum<T>> implements Argument {
+
+    private final T value;
+
+    public EnumArgument(T value) {
+        this.value = value;
+    }
+
+    public void apply(int position, PreparedStatement statement, StatementContext ctx) throws SQLException {
+        if (value == null) {
+            statement.setNull(position, Types.VARCHAR);
+        }
+        else {
+            statement.setString(position, value.name());
+        }
+    }
+
+    @Override
+    public String toString() {
+        return value != null ? value.toString() : "<null>";
+    }
+}
